# openapi.yaml
openapi: 3.0.3
info:
  title: AlfTekPro Multi-Tenant HRMS API
  version: 1.0.0
  description: |
    Multi-tenant HRMS system with row-level isolation, supporting multiple regions (UAE, USA, India).

    ## Architecture
    - **Authentication**: JWT Bearer tokens with tenant_id in claims
    - **Multi-Tenancy**: Row-level isolation with automatic query filtering
    - **Regions**: UAE (RTL/Arabic), USA (LTR/English), India (LTR/Hindi)

    ## Key Features
    - Employee Management with Temporal Job History
    - Shift & Roster Management
    - Attendance Tracking with Geofencing
    - Leave Management with Approval Workflows
    - Unified Action Center for Approvals
    - Payroll Processing with Regional Logic
    - Asset Management
    - Dynamic Forms (Region-Specific Fields)

  contact:
    name: AlfTekPro Support
    email: support@alftekpro.com
  license:
    name: Proprietary
servers:
  - url: https://api.alftekpro.com/v1
    description: Production Server
  - url: https://dev-api.alftekpro.com/v1
    description: Development Server
  - url: http://localhost:5001/v1
    description: Local Development
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      tags:
        - Employees
      summary: Get all employees
      description: Retrieve paginated list of employees with current job info
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Notice, Exited]
        - name: departmentId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search by name, email, or employee code
          schema:
            type: string
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/EmployeeListItem'
                      totalCount:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer

    post:
      tags:
        - Employees
      summary: Create new employee
      description: Create employee with initial job history record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      employeeId:
                        type: string
                        format: uuid
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /employees/{id}:
    get:
      tags:
        - Employees
      summary: Get employee details
      description: Retrieve employee with current job information and dynamic data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/EmployeeDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Employees
      summary: Update employee
      description: Update employee basic information (not job history)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /roster/{employeeId}:
    get:
      tags:
        - Roster
      summary: Get employee roster
      description: Retrieve shift assignments for employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Employee roster
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RosterEntry'

  # ==================== ATTENDANCE MODULE ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with claims: tenant_id, user_id, role

        Example header: Authorization: Bearer <token>
  parameters:
    PageNumber:
      name: pageNumber
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
...
    BadRequest:
      description: Bad Request - Validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
          description: i18n error code
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors keyed by field name

    # ==================== AUTH SCHEMAS ====================
...
    EmployeeListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        departmentName:
          type: string
          nullable: true
        designationTitle:
          type: string
          nullable: true
        status:
          type: string
          enum: [Active, Notice, Exited]
    EmployeeDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
        joiningDate:
          type: string
          format: date
        status:
          type: string
        currentJob:
          type: object
          nullable: true
          properties:
            departmentId:
              type: string
              format: uuid
            departmentName:
              type: string
            designationId:
              type: string
              format: uuid
            designationTitle:
              type: string
            reportingManagerId:
              type: string
              format: uuid
              nullable: true
            managerName:
              type: string
              nullable: true
            locationId:
              type: string
              format: uuid
              nullable: true
            salaryTierId:
              type: string
              format: uuid
              nullable: true
        dynamicData:
          type: object
          additionalProperties: true
          description: Region-specific fields (JSONB)
    CreateEmployeeRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - joiningDate
        - departmentId
        - designationId
      properties:
        employeeCode:
          type: string
          description: Auto-generated if not provided
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        joiningDate:
          type: string
          format: date
        departmentId:
          type: string
          format: uuid
        designationId:
          type: string
          format: uuid
        reportingManagerId:
          type: string
          format: uuid
          nullable: true
        locationId:
          type: string
          format: uuid
          nullable: true
        salaryTierId:
          type: string
          format: uuid
          nullable: true
        dynamicData:
          type: object
          additionalProperties: true
          description: Region-specific fields
    UpdateEmployeeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
        status:
          type: string
          enum: [Active, Notice, Exited]
        dynamicData:
          type: object
          additionalProperties: true
...
    RosterEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        shiftId:
          type: string
          format: uuid
        shiftName:
          type: string
        effectiveDate:
          type: string
          format: date

    # ==================== ATTENDANCE SCHEMAS ====================
...